+++ test/meetingsession/MeetingSessionStatus.test.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import * as chai from 'chai';
@@ -25,6 +25,7 @@ describe('MeetingSessionStatus', () => {
     MeetingSessionStatusCode.AudioInternalServerError,
     MeetingSessionStatusCode.AudioServiceUnavailable,
     MeetingSessionStatusCode.AudioDisconnected,
+    MeetingSessionStatusCode.ForbiddenToFetchTrunCredentials,
     MeetingSessionStatusCode.VideoCallSwitchToViewOnly,
     MeetingSessionStatusCode.VideoCallAtSourceCapacity,
     MeetingSessionStatusCode.SignalingBadRequest,
@@ -70,6 +71,7 @@ describe('MeetingSessionStatus', () => {
         MeetingSessionStatusCode.AudioCallEnded,
         MeetingSessionStatusCode.AudioInternalServerError,
         MeetingSessionStatusCode.AudioDisconnected,
+        MeetingSessionStatusCode.ForbiddenToFetchTrunCredentials,
         MeetingSessionStatusCode.SignalingBadRequest,
         MeetingSessionStatusCode.SignalingInternalServerError,
         MeetingSessionStatusCode.SignalingRequestFailed,