+++ test/task/ReceiveTURNCredentialsTask.test.ts
@@ -7,6 +7,7 @@ import AudioVideoControllerState from '../../src/audiovideocontroller/AudioVideo
 import NoOpAudioVideoController from '../../src/audiovideocontroller/NoOpAudioVideoController';
 import MeetingSessionConfiguration from '../../src/meetingsession/MeetingSessionConfiguration';
 import MeetingSessionCredentials from '../../src/meetingsession/MeetingSessionCredentials';
+import MeetingSessionStatusCode from '../../src/meetingsession/MeetingSessionStatusCode';
 import MeetingSessionTURNCredentials from '../../src/meetingsession/MeetingSessionTURNCredentials';
 import MeetingSessionURLs from '../../src/meetingsession/MeetingSessionURLs';
 import TimeoutScheduler from '../../src/scheduler/TimeoutScheduler';
@@ -74,6 +75,20 @@ describe('ReceiveTURNCredentialsTask', () => {
       task = new ReceiveTURNCredentialsTask(context);
       await task.run();
       expect(context.turnCredentials).to.equal(null);
+    });
+
+    it('handles a 403 error to fetch TURN Credentials', async () => {
+      domMockBehavior.fetchSucceeds = true;
+      domMockBehavior.responseStatusCode = 403;
+      domMockBehavior.responseSuccess = false;
+      try {
+        await task.run();
+        throw new Error('This line should not be reached.');
+      } catch (error) {
+        expect(error.message).includes(
+          `the meeting status code: ${MeetingSessionStatusCode.ForbiddenToFetchTrunCredentials}`
+        );
+      }
     });
   });